Fix truncated response in FakeWSRequestHolder
A `Result.body.dataStream` is a `Source[ByteString, _]` which generates
a stream of ByteString elements.

Using a `Sink.headOption` in runWith only processes the first (optional)
element and ignores all subsequent elements.

When using a MockWS with `Ok.sendFile` in a specification, the
controller using the `WSRequest` only received the first 8192 Bytes of
the given file because of this.

Concatenate all generated `ByteString`s and use that as responseBody.